+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 ### Improvements
 
 * fix(test): use lowercase instance name prefix in test suite
-* Bump Kubernetes SDK from 1.30.2 to 1.31.0
+* chore(deps): bump Kubernetes SDK from 1.30.2 to 1.31.0
+* chore(golang): bump golang from 1.22 to 1.23
 * chore(doc): add note about make for macOS users
 * go.mk: upgrade to v2.0.3 #89